I add <Cast /> component.
