+++ components/Cast/index.tsx
@@ -1,0 +1,3 @@
+import Cast from './Cast';
+
+export default Cast;
