+++ hooks/useFeed.tsx
@@ -8,7 +8,6 @@ const useFeed = () => {
     const init = async () => {
       const response = await getFeed('spotify.com/track');
       const soundCloud = await getFeed('soundcloud.com');
-      console.log('SWEETS SOUNDCLOUD', soundCloud);
       const combinedFeeds = [...response.casts, ...soundCloud.casts];
       const sortedFeeds = combinedFeeds.sort(
         (a, b) => new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime(),
